Move Functions to single recorded test class
