+++ tools/Azure.Mcp.Tools.Functions/tests/Azure.Mcp.Tools.Functions.Tests/assets.json
@@ -2,5 +2,5 @@
     "AssetsRepo": "Azure/azure-sdk-assets",
     "AssetsRepoPrefixPath": "",
     "TagPrefix": "Azure.Mcp.Tools.Functions.Tests",
-    "Tag": "Azure.Mcp.Tools.Functions.Tests_338e153b5b"
+    "Tag": "Azure.Mcp.Tools.Functions.Tests_95dc5fbc1d"
 }
